requirements.txt
# Core Dependencies
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface>=0.1.0
huggingface-hub>=0.24.0
sentence-transformers==3.2.1
langchain-groq>=0.2.0

# Vector Database
faiss-cpu==1.12.0

# Backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# Frontend
streamlit==1.39.0
requests==2.32.0

# PDF Processing
pypdf==5.1.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.0
